change HighlightToken byte offsets from u64 to i32

+++ shared/src/commonMain/kotlin/com/example/glance/Processing/HighlightToken.kt
@@ -1,7 +1,7 @@
 package com.example.glance.Processing
 
 data class HighlightToken(
-    val startByte: Long,
-    val endByte: Long,
+    val startByte: Int,
+    val endByte: Int,
     val kind: String,
 )
